+++ R/wrappers.R
@@ -8,32 +8,79 @@
 #'
 #' @param data a data frame that can be coerced into a \code{\link[tibble]{tibble}}
 #' @param formula a formula with the response variable on the left and the explanatory on the right
-#' @param alternative character string specifying the direction of the alternative hypothesis. Options are
+#' @param order #' @param order a string vector of specifying the order in which the levels of
+#' the explanatory variable should be ordered for subtraction, where
+#' \code{order = c("first", "second")} means \code{("first" - "second")}
+#' @param alternative character string giving the direction of the alternative hypothesis. Options are
 #' "\code{two_sided}" (default), "\code{greater}", or "\code{less}".
-#' @param ... currently ignored
+#' @param mu a numeric value giving the hypothesized null mean value for a one sample test
+#' and the hypothesized difference for a two sample test
+#' @param conf_int a logical value for whether to include the confidence interval or not. TRUE by default
+#' @param conf_level a numeric value between 0 and 1. Default value is 0.95 
+#' @param ... for passing in other arguments to \code{stats::t.test}
+#' @importFrom rlang f_lhs
+#' @importFrom rlang f_rhs
 #' @export
 #' @examples
 #' # t test for comparing mpg against automatic/manual
 #'   mtcars %>%
 #'     dplyr::mutate(am = factor(am)) %>%
-#'     t_test(mpg ~ am, alternative = "less")
+#'     t_test(mpg ~ am, order = c("1", "0"), alternative = "less")
 
 
 t_test <- function(data, formula, #response = NULL, explanatory = NULL,
-                   alternative = "two_sided",...){
-
+                   order = NULL,
+                   alternative = "two_sided", mu = 0, 
+                   conf_int = TRUE,
+                   conf_level = 0.95,
+                   ...){
+  
+  check_conf_level(conf_level)
+  
   # Match with old "dot" syntax
   if(alternative == "two_sided")
     alternative <- "two.sided"
 
   ### Only currently working with formula interface
 #  if (hasArg(formula)) {
-    data %>%
+  if(!is.null(f_rhs(formula))){
+    
+    data[[as.character(f_rhs(formula))]] <-
+      factor(data[[as.character(f_rhs(formula))]],
+             levels = c(order[1], order[2]))
+    
+    # Two sample case
+    prelim <- data %>%
       stats::t.test(formula = formula, data = .,
-                    alternative = alternative) %>%
-      broom::glance() %>%
+                    alternative = alternative,
+                    mu = mu, 
+                    conf.level = conf_level, ...) %>%
+      broom::glance()
+  } else {
+    # One sample case
+    # To fix weird indexing error convert back to data.frame
+    # (Error: Can't use matrix or array for column indexing)
+    data <- as.data.frame(data)
+    prelim <- stats::t.test(x = data[[as.character(f_lhs(formula))]],
+                  alternative = alternative,
+                  mu = mu, 
+                  conf.level = conf_level, ...) %>% 
+      broom::glance()
+  }
+  
+  if(conf_int){
+    results <- prelim %>% 
+      dplyr::select(statistic, t_df = parameter, p_value = p.value,
+                    alternative, 
+                    lower_ci = conf.low,
+                    upper_ci = conf.high)
+  } else {
+    results <- prelim %>% 
       dplyr::select(statistic, t_df = parameter, p_value = p.value,
                     alternative)
+  }
+    
+    return(results)
 #  } else {
     # data %>%
     #   stats::t.test(formula = substitute(response) ~ substitute(explanatory),
@@ -55,14 +102,13 @@ t_test <- function(data, formula, #response = NULL, explanatory = NULL,
 #'
 #' @param data a data frame that can be coerced into a \code{\link[tibble]{tibble}}
 #' @param formula a formula with the response variable on the left and the explanatory on the right
-#' @param ... currently ignored
+#' @param ... pass in arguments to {infer} functions
 #' @export
 
 t_stat <- function(data, formula, ...){
-  data %>%
-    t_test(formula = formula, ...) %>%
-    dplyr::select(statistic) %>%
-    dplyr::pull()
+  data %>% 
+    specify(formula = formula) %>% 
+    calculate(stat = "t", ...)
 }
 
 #'
@@ -82,6 +128,10 @@ t_stat <- function(data, formula, ...){
 chisq_test <- function(data, formula, #response = NULL, explanatory = NULL,
                        ...){
 
+  if(is.null(f_rhs(formula)))
+    stop(paste("`chisq_test()` currently only has functionality for",
+               "Chi-Square Test of Independence, not for Chi-Square",
+               "Goodness of Fit."))
   ## Only currently working with formula interface
   explanatory_var <- f_rhs(formula)
   response_var <- f_lhs(formula)
@@ -96,12 +146,28 @@ chisq_test <- function(data, formula, #response = NULL, explanatory = NULL,
 #'
 #' @param data a data frame that can be coerced into a \code{\link[tibble]{tibble}}
 #' @param formula a formula with the response variable on the left and the explanatory on the right
-#' @param ... additional arguments for \code{chisq.test}
+#' @param ... additional arguments for \code{stats::chisq.test}
 #' @export
 
 chisq_stat <- function(data, formula, ...){
-  data %>%
-    chisq_test(formula = formula, ...) %>%
-    dplyr::select(statistic) %>%
-    dplyr::pull()
+  
+  if(is.null(f_rhs(formula))){
+    stop(paste("`chisq_stat()` currently only has functionality for",
+               "Chi-Square Test of Independence, not for Chi-Square",
+               "Goodness of Fit. Use `specify() %>% hypothesize()",
+               "%>% calculate()` instead."))
+  } else {
+    data %>%
+      specify(formula = formula, ...) %>%
+      calculate(stat = "Chisq")
+  }
+}
+
+
+check_conf_level <- function(conf_level){
+  if(class(conf_level) != "numeric" |
+     conf_level < 0 |
+     conf_level > 1)
+    stop("The `conf_level` argument must be a number between 0 and 1.",
+         call. = FALSE)
 }