+++ inst/doc/chisq_test.R
@@ -23,10 +23,14 @@ fli_small <- flights %>%
          day_hour, origin, carrier)
 
 ## ------------------------------------------------------------------------
+obs_chisq <- fli_small %>%
+  specify(origin ~ season) %>% # alt: response = origin, explanatory = season
+  calculate(stat = "Chisq")
+
+## ------------------------------------------------------------------------
 obs_chisq <- fli_small %>% 
   chisq_test(formula = origin ~ season) %>% 
-  dplyr::select(statistic) %>% 
-  dplyr::pull()
+  dplyr::select(statistic)
 
 ## ------------------------------------------------------------------------
 obs_chisq <- fli_small %>% 
@@ -42,12 +46,11 @@ chisq_null_distn %>% visualize(obs_stat = obs_chisq, direction = "greater")
 
 ## ------------------------------------------------------------------------
 chisq_null_distn %>% 
-  dplyr::summarize(p_value = mean(stat >= obs_chisq)) %>% 
-  dplyr::pull()
+  get_pvalue(obs_stat = obs_chisq, direction = "greater")
 
 ## ------------------------------------------------------------------------
 fli_small %>%
-  specify(origin ~ season) %>% # alt: response = origin, explanatory = season
+  specify(origin ~ season) %>% 
   hypothesize(null = "independence") %>%
   # generate() ## Not used for theoretical
   calculate(stat = "Chisq") %>%
@@ -55,7 +58,7 @@ fli_small %>%
 
 ## ----eval=FALSE----------------------------------------------------------
 #  fli_small %>%
-#    specify(origin ~ season) %>% # alt: response = origin, explanatory = season
+#    specify(origin ~ season) %>%  %>% # alt: response = origin, explanatory = season
 #    hypothesize(null = "independence") %>%
 #    generate(reps = 1000, type = "permute") %>%
 #    calculate(stat = "Chisq") %>%